Added cpg-context tool
This tool parses a cpg format network and outputs a json formatted analysis of
context availability in different scopes of the file. If an error occurs, it outputs the
json formatted error message. This information can be used in a text editor (using
a plugin) to show useful information while editing.