Implemented property constraints
Property constraints ensure that the value of a property is always validated by its constraint expression.

In the cpg-format, constraints can be specified in <>:
state "s"
{
	x = 0 <cycle(this, 0, 2 * pi)>
}

A new special keyword in expressions "this" can be used to refer to the containing property from an expression. Two new mathematical functions are added that can be used for often used constraints, clip and cycle.