Make expansion context a stack
This is a large change in terms of how defines and expansions are kept
in context (selections, parsing). Whereas before, every context was
duplicated (although sometimes smartly copy-on-write) in selections
as well as kept in a separate CdnEmbeddedContext, all contexts are
now handled by CdnExpansionContext. This new expansion context
is a stack of expansions and defines which greatly simplifies
maintaining it, but also should be much faster to use since defines
and expansions never have to be copied anymore.